# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: dLNk Attack Platform API
  description: |
    Comprehensive API documentation for dLNk Attack Platform.
    
    ## Features
    - Authentication & Authorization
    - Target Management
    - Attack Orchestration
    - Agent Control
    - Zero-Day Hunter
    - Real-time Updates
    
    ## Authentication
    All endpoints require JWT token authentication except `/auth/login` and `/auth/register`.
    Include the token in the `Authorization` header: `Bearer <token>`
    
  version: 1.0.0
  contact:
    name: dLNk Security Team
    email: security@dlnk.io
  license:
    name: Proprietary
    url: https://dlnk.io/license
servers:
  - url: https://api.dlnk.io/v1
    description: Production server
  - url: https://staging-api.dlnk.io/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
              properties:
                username:
                  type: string
                  example: newuser
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input or user already exists
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
...